backend/pyproject.toml
[project]
name = "meeting-ai"
version = "0.6.0"
description = "离线会议纪要 AI：语音转写、说话人分离、智能命名、会议总结"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "Your Name", email = "you@example.com" }]
keywords = ["meeting", "transcription", "diarization", "whisper", "offline", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # 核心依赖（阶段 1-3）
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "rich>=13.0.0",           # 美观的终端输出
    "typer>=0.9.0",           # CLI 框架
    
    # 音频处理
    "soundfile>=0.12.0",
    "librosa>=0.10.0",
    
    # ASR（阶段 3）
    "faster-whisper>=1.0.0",
    
    # 说话人分离（阶段 1-2）
    "pyannote-audio>=3.1.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    
    # LLM（阶段 4）
    "llama-cpp-python>=0.2.0",

    # API 后端（阶段 7）
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "python-multipart>=0.0.6",
    "websockets>=11.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",            # 代码检查和格式化
    "mypy>=1.0.0",            # 类型检查
    "pre-commit>=3.0.0",
]
gui = [
    # 桌面 GUI（阶段 5）
    "flet>=0.21.0",
]
enhance = [
    # 音频增强
    "noisereduce>=3.0.0",
]
stream = [
    # 实时流式（阶段 6）
    "websockets>=11.0.0",
    "sounddevice>=0.4.6",
    "webrtcvad>=2.0.10",
    "funasr>=1.0.0",          # 阿里 FunASR 流式语音识别
]

[project.scripts]
meeting-ai = "meeting_ai.cli:app"
meeting-ai-api = "meeting_ai.api.main:app"

[project.gui-scripts]
meeting-ai-gui = "meeting_ai.gui:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/meeting_ai"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]  # 行长度由 line-length 控制

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=meeting_ai --cov-report=term-missing"
